tests, move analytical boson sampling fns to their own file, add typehints
