+++ open_qbench/fidelities.py
@@ -6,6 +6,8 @@ import math
 import numpy as np
 from ptseries.tbi import create_tbi
 
+from .boson_sampling import num_possible_samples
+
 
 def normalized_fidelity(dist_ideal: dict, dist_backend: dict) -> float:
     """Normalized fidelity of Lubinski et al."""
@@ -17,15 +19,12 @@ def normalized_fidelity(dist_ideal: dict, dist_backend: dict) -> float:
     return fidelity
 
 
-def create_normalized_fidelity(input_state):
-    def normalized_fidelity_orca(dist_ideal: dict, dist_backend: dict) -> float:
+def create_normalized_fidelity(input_state: tuple):
+    """Create a normalized fidelity function for a given input state -> ORCA PT"""
+    def normalized_fidelity_orca(dist_ideal: dict[tuple, float], dist_backend: dict[tuple, float]) -> float:
         """Normalized fidelity modified for Boson Sampling"""
-        total_photons = np.sum(input_state)
-        num_bins = len(input_state)
-        num_possible = 0
-        for tp in range(total_photons+1):
-            num_possible += math.comb(tp+num_bins-1, num_bins-1)  # num of possible correct outputs
-
+        total_photons = sum(input_state)
+        num_possible = num_possible_samples(input_state)
         # filter out only 'correct' samples
         dist_ideal = {k: v for k, v in dist_ideal.items() if np.sum(k) <= total_photons}
         dist_backend = {k: v for k, v in dist_backend.items() if np.sum(k) <= total_photons}
@@ -72,139 +71,14 @@ def fidelity_with_uniform(dist: dict, num_possible_samples: int | None = None) -
         float: Classical fidelity given by:
         F(X,Y) = (\sum _i \sqrt{p_i q_i})^2
     """
-    num_qubits = len(next(iter(dist.keys())))
+
     fidelity = 0
     if num_possible_samples is None:
+        num_qubits = len(next(iter(dist.keys())))
         num_possible_samples = 2**num_qubits
+
     uniform_prob = 1 / num_possible_samples
     for prob in dist.values():
         fidelity += math.sqrt(prob * uniform_prob)
     fidelity = fidelity**2
     return fidelity
-
-
-def generate_all_samples_orca(input_state):
-    time_bin_interferometer = create_tbi()
-    samples = time_bin_interferometer.sample(
-        input_state=input_state,
-        theta_list=[np.pi / 4] * (len(input_state) - 1),  # 50/50 beam splitters
-        n_samples=100000,
-    )
-    samples_sorted = dict(sorted(samples.items(), key=lambda state: -state[1]))
-    labels = list(samples_sorted.keys())
-    return labels
-
-
-def generate_all_possible_outputs_orca(input_state):
-    all_possible_outputs = []
-    max_num_photons = np.sum(input_state)
-    # print(max_num_photons)
-    tab_num_photons = range(0, max_num_photons + 1)
-    arr = itertools.product(tab_num_photons, repeat=len(input_state))
-
-    for el in arr:
-        if np.sum(el) <= np.sum(input_state):
-            all_possible_outputs.append(el)
-    return all_possible_outputs
-
-
-def per(mtx, column, selected, prod, output=False):
-    if column == mtx.shape[1]:
-        if output:
-            print(selected, prod)
-        return prod
-    else:
-        result = 0
-        for row in range(mtx.shape[0]):
-            if row not in selected:
-                result = result + per(
-                    mtx,
-                    column + 1,
-                    [*selected, row],
-                    prod * mtx[row, column],
-                )
-        return result
-
-
-def compute_permanent(mat):
-    return per(mat, 0, [], 1)
-
-
-def construct_unitary(bm_parameters):
-    def beamsplitter_matrix(theta, n, i, j):
-        U = np.eye(n, dtype=complex)
-        cos_theta = np.cos(theta)
-        sin_theta = np.sin(theta)
-
-        U[i, i] = cos_theta
-        U[i, j] = -sin_theta
-        U[j, i] = sin_theta
-        U[j, j] = cos_theta
-
-        return U
-
-    def apply_sequential_beam_splitters(thetas):
-        n = len(thetas) + 1
-        U = np.eye(n, dtype=complex)
-        for k, theta in enumerate(thetas):
-            if k + 1 < n:
-                BS = beamsplitter_matrix(theta, n, k, k + 1)
-                U = BS @ U
-        return U
-
-    return apply_sequential_beam_splitters(bm_parameters)
-
-
-def generate_submatrix(U, input_bitstring, output):
-    columns = [j for j, t in enumerate(output) for _ in range(t)]
-    UT = U[:, columns]
-
-    rows = [i for i, s in enumerate(input_bitstring) for _ in range(s)]
-    UST = UT[rows, :]
-    return UST
-
-
-def output_probabilities(input_bitstring, U):
-    possible_outputs = generate_all_possible_outputs_orca(input_bitstring)
-    probabilities = {}
-
-    for output_indices in possible_outputs:
-        U_submatrix = generate_submatrix(U, input_bitstring, output_indices)
-        PQT = np.abs(compute_permanent(U_submatrix)) ** 2
-        probabilities[tuple(output_indices)] = PQT / (
-            np.prod([math.factorial(s) for s in input_bitstring])
-            * np.prod([math.factorial(t) for t in output_indices])
-        )
-
-    total_prob = sum(probabilities.values())
-    for key in probabilities:
-        probabilities[key] /= total_prob
-
-    return probabilities
-
-
-def generate_analitically(input_bitstring):
-    input_bitstring = input_bitstring[::-1]
-    theta_list = [np.pi / 4] * (len(input_bitstring) - 1)
-    U = construct_unitary(theta_list)
-    probabilities = output_probabilities(input_bitstring, U)
-    return sorted(
-        [
-            (key[::-1])
-            for key, value in probabilities.items()
-            if value > 1e-30 and sum(key) == sum(input_string)
-        ]
-    )
-
-
-if __name__ == "__main__":
-    input_string = (1, 1, 2)
-    ans0 = generate_analitically(input_string)
-    ans1 = generate_all_samples_orca(input_string)
-    ans2 = generate_all_possible_outputs_orca(input_string)
-    print(ans0)
-    print(len(ans0))
-    print(ans1)
-    print(len(ans1))
-    print(ans2)
-    print(len(ans2))
